Renames upload to s3
