+++ gulpfile.babel.js
@@ -14,7 +14,7 @@ import bump from './gulp/bump';
 import * as clean from './gulp/clean';
 import * as rsx from './gulp/rsx';
 import * as revision from './gulp/revision';
-import * as upload from './gulp/upload';
+import * as uploadToS3 from './gulp/s3';
 import * as git from './gulp/git';
 import * as docker from './gulp/docker';
 import * as localization from './gulp/localization';
@@ -59,11 +59,11 @@ gulp.task('rsx:build_urls', rsx.buildUrls);
 gulp.task('rsx:codex_urls', rsx.codexUrls);
 gulp.task('revision:generate', revision.generate);
 gulp.task('revision:replace', revision.replace);
-gulp.task('upload:main', upload.main);
-gulp.task('upload:audio', upload.audio);
-gulp.task('upload:main:versioned', () => upload.main(version));
-gulp.task('upload:audio:versioned', () => upload.audio(version));
-gulp.task('upload:stagingcdn', upload.stagingcdn);
+gulp.task('upload:main', uploadToS3.main);
+gulp.task('upload:audio', uploadToS3.audio);
+gulp.task('upload:main:versioned', () => uploadToS3.main(version));
+gulp.task('upload:audio:versioned', () => uploadToS3.audio(version));
+gulp.task('upload:stagingcdn', uploadToS3.stagingcdn);
 gulp.task('changelog', git.changelog);
 gulp.task('docker:build', docker.build);
 gulp.task('docker:tag', docker.tag);
